Cleaned up azure.cli.commands to make it clear what is supposed to be part of the public surface area for command authors and what is internal.
Changed the command package protocol back to using a single command table defined in the azure.cli.commands packge instead of expecting a get_command_table method for each command package.

Moved argument aliasing and updating logic from application into azure.cli.commands since that is where it belongs.
