+++ src/command_modules/azure-cli-component/azure/cli/command_modules/component/__init__.py
@@ -1,3 +1,5 @@
 # pylint: disable=unused-import
-from azure.cli.command_modules.component.generated import command_table
+
 import azure.cli.command_modules.component._params
+import azure.cli.command_modules.component.generated
+import azure.cli.command_modules.component.custom
