+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/__init__.py
@@ -1,3 +1,5 @@
 # pylint: disable=unused-import
-from azure.cli.command_modules.network.generated import command_table
+
 import azure.cli.command_modules.network._params
+import azure.cli.command_modules.network.generated
+import azure.cli.command_modules.network.custom
