+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/_params.py
@@ -1,16 +1,11 @@
 # pylint: disable=line-too-long
 import argparse
 
-from azure.mgmt.network import NetworkManagementClient, NetworkManagementClientConfiguration
-
 from azure.cli.command_modules.network._actions import LBDNSNameAction
-from azure.cli.commands._command_creation import get_mgmt_service_client
-from azure.cli.commands._params import location_type
-from azure.cli.commands.argument_types import register_cli_argument, CliArgumentType
+from azure.cli.commands.parameters import location_type
+from azure.cli.commands import register_cli_argument, CliArgumentType
 # FACTORIES
 
-def _network_client_factory(**_):
-    return get_mgmt_service_client(NetworkManagementClient, NetworkManagementClientConfiguration)
 
 # BASIC PARAMETER CONFIGURATION
