+++ src/command_modules/azure-cli-profile/azure/cli/command_modules/profile/__init__.py
@@ -1,3 +1,5 @@
 # pylint: disable=unused-import
-from azure.cli.command_modules.profile.generated import command_table
+
 import azure.cli.command_modules.profile._params
+import azure.cli.command_modules.profile.generated
+import azure.cli.command_modules.profile.custom
