+++ src/command_modules/azure-cli-profile/azure/cli/command_modules/profile/_params.py
@@ -1,4 +1,4 @@
-from azure.cli.commands.argument_types import CliArgumentType, register_cli_argument
+from azure.cli.commands import CliArgumentType, register_cli_argument
 
 # BASIC PARAMETER CONFIGURATION
