+++ src/command_modules/azure-cli-profile/azure/cli/command_modules/profile/generated.py
@@ -1,11 +1,8 @@
 from __future__ import print_function
 
-from azure.cli.commands import CommandTable
-from azure.cli.commands.command_types import cli_command
+from azure.cli.commands import command_table, cli_command
 
 from .custom import login, logout, list_subscriptions, set_active_subscription, account_clear
-
-command_table = CommandTable()
 
 cli_command(command_table, 'login', login)
 cli_command(command_table, 'logout', logout)
