+++ src/command_modules/azure-cli-resource/azure/cli/command_modules/resource/__init__.py
@@ -1,3 +1,5 @@
 # pylint: disable=unused-import
-from azure.cli.command_modules.resource.generated import command_table
+
 import azure.cli.command_modules.resource._params
+import azure.cli.command_modules.resource.generated
+import azure.cli.command_modules.resource.custom
