+++ src/command_modules/azure-cli-resource/azure/cli/command_modules/resource/_factory.py
@@ -1,6 +1,6 @@
 from azure.mgmt.resource.resources import (ResourceManagementClient,
                                            ResourceManagementClientConfiguration)
-from azure.cli.commands._command_creation import get_mgmt_service_client
+from azure.cli.commands.client_factory import get_mgmt_service_client
 
 def _resource_client_factory(**_):
     return get_mgmt_service_client(ResourceManagementClient, ResourceManagementClientConfiguration)
