+++ src/command_modules/azure-cli-resource/azure/cli/command_modules/resource/_params.py
@@ -1,8 +1,8 @@
 # pylint: disable=line-too-long
 import argparse
 
-from azure.cli.commands._params import resource_group_name_type, tag_type, tags_type
-from azure.cli.commands.argument_types import register_cli_argument, CliArgumentType
+from azure.cli.commands import register_cli_argument, CliArgumentType
+from azure.cli.commands.parameters import resource_group_name_type, tag_type, tags_type
 
 from ._validators import validate_resource_type, validate_parent, resolve_resource_parameters
