+++ src/command_modules/azure-cli-resource/azure/cli/command_modules/resource/tests/test_validators.py
@@ -1,7 +1,7 @@
 import unittest
 from six import StringIO
 
-from azure.cli.commands._validators import *
+from azure.cli.commands.validators import *
 
 from azure.cli.command_modules.resource._validators import *
