+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/__init__.py
@@ -1,3 +1,4 @@
 # pylint: disable=unused-import
-from azure.cli.command_modules.role.generated import command_table
+
+import azure.cli.command_modules.role.generated
 import azure.cli.command_modules.role._params
