--- src/command_modules/azure-cli-role/azure/cli/command_modules/role/custom.py
@@ -1,5 +1,0 @@
-# pylint: disable=too-few-public-methods
-import azure.cli._logging as _logging
-
-logger = _logging.get_az_logger(__name__)
-
