+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/__init__.py
@@ -1,3 +1,5 @@
 # pylint: disable=unused-import
-from azure.cli.command_modules.storage.generated import command_table
+
 import azure.cli.command_modules.storage._params
+import azure.cli.command_modules.storage.generated
+import azure.cli.command_modules.storage.custom
