+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/_factory.py
@@ -5,7 +5,7 @@ from azure.storage.blob import BlockBlobService
 from azure.storage.file import FileService
 from azure.storage._error import _ERROR_STORAGE_MISSING_INFO
 
-from azure.cli.commands._command_creation import get_mgmt_service_client, get_data_service_client
+from azure.cli.commands.client_factory import get_mgmt_service_client, get_data_service_client
 from azure.cli.commands import CLIError
 from ._params import blob_types
