+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/__init__.py
@@ -1,4 +1,6 @@
 # pylint: disable=unused-import
-from azure.cli.command_modules.vm.generated import command_table
+
+import azure.cli.command_modules.vm._help
 import azure.cli.command_modules.vm._params
-from ._help import helps
+import azure.cli.command_modules.vm.generated
+import azure.cli.command_modules.vm.custom
