+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_factory.py
@@ -1,5 +1,5 @@
-from azure.cli.commands._command_creation import (get_mgmt_service_client,
-                                                  get_subscription_service_client)
+from azure.cli.commands.client_factory import (get_mgmt_service_client,
+                                               get_subscription_service_client)
 from azure.mgmt.compute import ComputeManagementClient, ComputeManagementClientConfiguration
 
 def _compute_client_factory(**_):
