+++ src/command_modules/azure-cli-webapp/azure/cli/command_modules/webapp/__init__.py
@@ -1,3 +1,4 @@
 # pylint: disable=unused-import
-from azure.cli.command_modules.webapp.generated import command_table
+
 import azure.cli.command_modules.webapp._params
+import azure.cli.command_modules.webapp.generated
